JOY-101 refactor: MyPayments, updated styles
