+++ src/features/post/ui/createPost/steps/stepDescription/StepDescription.tsx
@@ -4,9 +4,9 @@ import { ChangeEvent } from 'react'
 
 import { usePostContext } from '@/features/post/ui/createPost/providers'
 import { ImageSlider } from '@/shared/ui/imageSlider'
+import { User, UserCard } from '@/entities/user'
 
-import { User } from '@/entities/user/ui/userCard'
-import { PublicationDescription, UserCard } from '@/shared/ui'
+import { PublicationDescription } from '@/shared/ui'
 import s from './StepDescription.module.scss'
 
 type StepDescriptionProps = {
