+++ src/features/profile/ui/myPayments/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './MyPayments'
