+++ src/features/profile/ui/myPayments/ui/paymentsTable/index.ts
