+++ src/features/profile/ui/profileSettings/ProfileSettings.module.scss
@@ -2,7 +2,6 @@
   width: 100%;
   max-width: 972px;
   margin: 0 auto;
-  padding: 20px;
 }
 
 .comingSoon {
