+++ src/features/profile/ui/profileSettings/ProfileSettings.tsx
@@ -7,9 +7,9 @@ import { Tabs } from '@/shared/ui'
 import { Tab } from '@/shared/ui/tabs'
 import { GeneralInformation } from '../generalInformation/GeneralInformation'
 import { PATH } from '@/shared/config/routes'
+import { MyPayments } from '../myPayments'
 
 import s from './ProfileSettings.module.scss'
-import { MyPayments } from '../myPayments/MyPayments'
 
 type ProfileSettingsProps = {
   activePart: string
