+++ src/shared/ui/index.ts
@@ -15,7 +15,6 @@ export { DatePicker } from './datePicker'
 export { DateStamp } from './dateStamp'
 export { DropdownMenu, DropdownMenuItem } from './dropdownMenu'
 
-export { createSidebarItems, Sidebar } from '../../widgets/sidebar'
 export { Form } from './form'
 export { ImageSlider } from './imageSlider'
 export { Loader } from './loader'
@@ -31,3 +30,4 @@ export { Tabs } from './tabs'
 export { TextArea } from './textArea'
 export { TextField } from './textField'
 export { Typography } from './typography'
+export { ToastSnackbar } from './toastSnackbar'
