+++ src/shared/ui/toastSnackbar/index.ts
@@ -1,0 +1,1 @@
+export * from './ToastSnackbar'
